fix(erp): 修复产品表 unit_id 字段类型并新增 PostgreSQL 支持
- 修复 erp_product 表的 unit_id 字段类型错误（INT → BIGINT）
- 新增完整的 PostgreSQL 数据库初始化脚本（1,644 行）
- 实现 MySQL → PostgreSQL 的自动类型转换
- 添加 33 个表的 BEFORE UPDATE 触发器以自动更新 update_time
- 完成实体类与数据库脚本的 100% 一致性验证

BREAKING CHANGE: 如果已部署使用 INT 类型的 unit_id，需要执行迁移 SQL：
ALTER TABLE erp_product MODIFY COLUMN unit_id BIGINT NOT NULL;

+++ yudao-module-erp/sql/erp-2024-05-03.sql
@@ -212,7 +212,7 @@ CREATE TABLE `erp_product`  (
   `name` varchar(100) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL COMMENT '产品名称',
   `bar_code` varchar(50) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL COMMENT '产品条码',
   `category_id` bigint NOT NULL COMMENT '产品分类编号',
-  `unit_id` int NOT NULL COMMENT '单位编号',
+  `unit_id` bigint NOT NULL COMMENT '单位编号',
   `status` tinyint NOT NULL COMMENT '产品状态',
   `standard` varchar(100) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NULL DEFAULT NULL COMMENT '产品规格',
   `remark` varchar(500) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NULL DEFAULT NULL COMMENT '产品备注',
